+++ Test/Unit/Model/QliroOrder/Builder/ValidateOrderBuilderStockTest.php
@@ -10,7 +10,10 @@ namespace Qliro\QliroOne\Test\Unit\Model\QliroOrder\Builder;
 use Magento\Quote\Model\CustomerManagement;
 use Magento\Quote\Model\Quote;
 use Magento\Quote\Model\Quote\Item as QuoteItem;
+use Magento\Quote\Api\CartRepositoryInterface;
 use Magento\Quote\Model\SubmitQuoteValidator;
+use Magento\Store\Model\App\Emulation as StoreEmulation;
+use Magento\Store\Model\StoreManagerInterface;
 use Magento\Store\Model\Store;
 use PHPUnit\Framework\TestCase;
 use Qliro\QliroOne\Api\Data\ValidateOrderNotificationInterface;
@@ -143,7 +146,10 @@ class ValidateOrderBuilderStockTest extends TestCase
             $this->createMock(LogManager::class),
             $this->createMock(SubmitQuoteValidator::class),
             $this->createMock(CustomerManagement::class),
-            $this->createMock(Config::class)
+            $this->createMock(Config::class),
+            $this->createMock(CartRepositoryInterface::class),
+            $this->createMock(StoreManagerInterface::class),
+            $this->createMock(StoreEmulation::class)
         );
 
         $builder->setQuote($this->buildQuote($quoteItems, $isActive));